server: idempotent style client
the client doesn't need to control the server process anymore. If there
is something running on 2442, it assumes it's the jsoncomma server. If
not, it starts it, and forgets about the process. The reason is that we
cannot reliably shutdown the process because sublime text doesn't have a
on_exit event